changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11754 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/Views/ScriptView/ORLineMarker.h
@@ -7,7 +7,7 @@
 }
 
 - (id) initWithRulerView:(NSRulerView *)aRulerView lineNumber:(float)line image:(NSImage *)anImage imageOrigin:(NSPoint)imageOrigin;
-- (void) setLineNumber:(unsigned)line;
-- (unsigned) lineNumber;
+- (void) setLineNumber:(NSUInteger)line;
+- (NSUInteger) lineNumber;
 
 @end
